+++ generate.py
@@ -1,5 +1,5 @@
 import torch
-from training import load_model, load_tokenizer, get_loaders, decode_ids, compute_text_metrics, setup_device, block_size
+# from training import load_model, load_tokenizer, get_loaders, decode_ids, compute_text_metrics, setup_device, block_size
 import torch.distributed as dist
 import os
 import torch.nn.functional as F
@@ -7,6 +7,7 @@ from training import get_loaders
 from torch.nn.parallel import DistributedDataParallel as DDP
 import argparse
 from utils.model_utils import *
+from training import load_model
 
 # Hyperparameters
 batch_size = 64
